# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FutureMap AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Single issue / signal detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: Global signal stream of issues, filterable by planet, category and country
      parameters:
        - in: query
          name: planet
          required: false
          schema:
            $ref: "#/components/schemas/Planet"
        - in: query
          name: category
          required: false
          schema:
            type: string
            enum: [news, conflict, disease, politics, economy, culture, ai_jobs, tech, natural_disaster, cyber, terror, climate, space, lunar_base, mars_habitat]
        - in: query
          name: country
          required: false
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
  /issues/summary:
    get:
      operationId: getIssueSummary
      tags: [issues, dashboard]
      summary: Counts of issues per category for the global signal stream
      parameters:
        - in: query
          name: planet
          required: false
          schema:
            $ref: "#/components/schemas/Planet"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"
  /jobs/analyze:
    post:
      operationId: analyzeJob
      tags: [jobs]
      summary: Generate an AI future report for a specified job and country
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobAnalyzeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobReport"
  /jobs/reports/recent:
    get:
      operationId: listRecentJobReports
      tags: [jobs, dashboard]
      summary: Recent job future reports generated across all users
      parameters:
        - in: query
          name: planet
          required: false
          schema:
            $ref: "#/components/schemas/Planet"
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 8 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobReport"
...
  /planets/{planet}/locations/{code}:
    get:
      operationId: getPlanetLocation
      tags: [planets]
      summary: Detail for a non-Earth planet location with related signals
      parameters:
        - in: path
          name: planet
          required: true
          schema:
            $ref: "#/components/schemas/Planet"
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanetLocationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Issue:
      type: object
      required: [id, countryCode, countryFlag, category, headline, planet, publishedAt]
      properties:
        id: { type: string }
        countryCode: { type: string }
        countryFlag: { type: string }
        cityId: { type: ["string", "null"] }
        planet: { $ref: "#/components/schemas/Planet" }
        category:
          type: string
          enum: [news, conflict, disease, politics, economy, culture, ai_jobs, tech, natural_disaster, cyber, terror, climate, space, lunar_base, mars_habitat]
        headline: { type: string }
        body: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        publishedAt: { type: string, format: date-time }
...
    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
          enum: [news, conflict, disease, politics, economy, culture, ai_jobs, tech, natural_disaster, cyber, terror, climate, space, lunar_base, mars_habitat]
        count: { type: integer, minimum: 0 }
    Planet:
      type: string
      enum: [earth, moon, mars]
    PlanetLocation:
      type: object
      required: [code, planet, name, nameKo, flag, latitude, longitude, descriptionKo, description]
      properties:
        code: { type: string }
        planet: { $ref: "#/components/schemas/Planet" }
        name: { type: string }
        nameKo: { type: string }
        flag: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        description: { type: string }
        descriptionKo: { type: string }
...
    PlanetLocationDetail:
      type: object
      required: [location, signals]
      properties:
        location: { $ref: "#/components/schemas/PlanetLocation" }
        signals:
          type: array
          items: { $ref: "#/components/schemas/Issue" }
    JobAnalyzeInput:
      type: object
      required: [jobName, countryCode]
      properties:
        jobName: { type: string, minLength: 1 }
        countryCode: { type: string, minLength: 2 }
        planet: { $ref: "#/components/schemas/Planet" }
    JobReport:
      type: object
      required:
        - id
        - jobName
        - countryCode
        - countryName
        - planet
        - automationRisk
        - growthScore
        - summary
        - automatedTasks
        - humanStrengths
        - futureChanges
        - recommendedSkills
        - countryOpportunities
        - createdAt
      properties:
        id: { type: string }
        jobName: { type: string }
        countryCode: { type: string }
        countryName: { type: string }
        planet: { $ref: "#/components/schemas/Planet" }
        automationRisk: { type: integer, minimum: 0, maximum: 100 }
        growthScore: { type: integer, minimum: 0, maximum: 100 }
        summary: { type: string }
        automatedTasks:
          type: array
          items: { type: string }
        humanStrengths:
          type: array
          items: { type: string }
        futureChanges:
          type: array
          items: { type: string }
        recommendedSkills:
          type: array
          items: { type: string }
        countryOpportunities:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...