# lib/api-zod/src/generated/types/planetLocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FutureMap AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Planet } from "./planet";

export interface PlanetLocation {
  code: string;
  planet: Planet;
  name: string;
  nameKo: string;
  flag: string;
  latitude: number;
  longitude: number;
  description: string;
  descriptionKo: string;
}